requirements.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
google-generativeai>=0.3.0
streamlit-mic-recorder>=0.0.5
soundfile>=0.12.1
python-dotenv